Add files via upload
The zip-folder contains a folder called sandbox. Inside the sandbox, there is a  pdf file with a report on runtimes. the file called lastAssessment.js is where I measure the performance of the two functions (doublerAppend and doublerInsert), while the file called lastAssessmentStepTwo.js contains the code of the functions given in the task. Could you pelase see eth code and advise fi any corrections need to be made? Thank you in advance.